Move truncate text logic into hooks file instead of component body
