+++ src/shared-components/alert/index.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useRef } from 'react';
+import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { useTheme } from 'emotion-theming';
 
@@ -6,6 +6,7 @@ import { CheckmarkIcon, ChevronIcon, ErrorIcon, InfoIcon } from '../../icons';
 import { Avatar } from '../avatar';
 import Style from './style';
 import { isDefined } from '../../utils/isDefined';
+import { useTruncateText } from './hooks';
 
 const ANIMATION_DELAY = 500;
 
@@ -63,7 +64,7 @@ export const Alert = (alertProps: AlertProps) => {
   const [exiting, setExiting] = useState(false);
   const [exited, setExited] = useState(false);
 
-  const contentText = useRef<HTMLDivElement>(null);
+  const { contentText } = useTruncateText(truncateText);
 
   let timer: number | undefined;
 
@@ -79,39 +80,6 @@ export const Alert = (alertProps: AlertProps) => {
       onExit(otherProps);
     }, ANIMATION_DELAY);
   };
-
-  useEffect(() => {
-    if (truncateText) {
-      const contentElement = contentText.current;
-      if (contentElement) {
-        const initialWords = contentElement.innerHTML;
-        const wordsArray = initialWords.split(' ');
-        /**
-         * TODO: Find better contentValues values for the logic.
-         * Secondary Theme truncated content gets stuck in the while loop because
-         * there are different dimensions at play.
-         */
-        const ARBITRARY_SECONDARY_OFFSET = 5;
-        while (
-          contentElement.scrollHeight >
-            contentElement.offsetHeight + ARBITRARY_SECONDARY_OFFSET &&
-          wordsArray.length !== 0
-        ) {
-          wordsArray.pop();
-          contentElement.innerHTML = `${wordsArray.join(' ')}...`;
-        }
-
-        /**
-         * If while loop pops all words due to element dimensions,
-         * prefer resetting without truncating to potentially
-         * broken functionality
-         */
-        if (wordsArray.length === 0) {
-          contentElement.innerHTML = initialWords;
-        }
-      }
-    }
-  });
 
   /**
    * Duration logic effect
